fix: ajusta model e migration para que número de registro seja um BigInt

+++ backend/src/app/database/ORM/migrations/06-User.ts
@@ -21,7 +21,7 @@ export default {
         field: migrations.columnName.passwordHash,
         allowNull: false,
       },
-      registry: { type: DataTypes.INTEGER, allowNull: false },
+      registry: { type: DataTypes.BIGINT, allowNull: false },
       name: {
         type: DataTypes.STRING,
         allowNull: false,
